refactor(LC006): split cartesian-explosion analysis helpers
- separate include-chain and navigation analysis
- keep analyzer behavior unchanged while reducing local complexity
